In the process of creating a simple UI to control the monitoring and Debugging
